% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.R
\name{conv}
\alias{conv}
\alias{ovt_to_yt}
\alias{vat_to_yt}
\alias{yt_to_vat}
\alias{yt_to_ovt}
\title{Convert between different ID numbers}
\usage{
ovt_to_yt(x)

vat_to_yt(x)

yt_to_vat(x)

yt_to_ovt(x)
}
\arguments{
\item{x}{The input string(s).}
}
\description{
Currently allows conversion from E-invoicing address and VAT number to
Business ID and vice versa.
}
