% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.R
\name{gen}
\alias{gen}
\alias{gen_yt}
\alias{gen_vat}
\alias{gen_ovt}
\alias{gen_pic}
\title{Generate valid examples of the strings}
\usage{
gen_yt(n = 1)

gen_vat(n = 1)

gen_ovt(n = 1)

gen_pic(n = 1)
}
\arguments{
\item{n}{The number of examples to be generated. Defaults to 1.}
}
\description{
Rules for generating the IDs obtained from
\url{http://tarkistusmerkit.teppovuori.fi/tarkmerk.htm}
\describe{
 \item{\code{gen_yt}}{generates a Business ID (y-tunnus).}
 \item{\code{gen_vat}}{generates a (Finnish) VAT Number (alv-tunnus).}
 \item{\code{gen_ovt}}{generates a (Finnish) OVT-code (ovt-tunnus).}
 \item{\code{gen_pic}}{generates a (Finnish) personal identification code (hetu).}
}
}
