% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\alias{valid_yt}
\alias{valid_ovt}
\alias{valid_vat}
\alias{valid_id}
\title{Check if the string is a valid ID Number}
\usage{
valid_yt(x, locate = TRUE, require_checksum = TRUE)

valid_ovt(x, locate = TRUE, require_checksum = TRUE)

valid_vat(x, locate = TRUE, require_checksum = TRUE)

valid_id(x, locate = TRUE, require_checksum = TRUE)
}
\arguments{
\item{x}{The input string(s).}

\item{locate}{If \code{TRUE} (default) tries to locate the valid part of
the string.}

\item{require_checksum}{If \code{TRUE} (default), require that the checksum
digit matches in addition to the format of the string. Note, does not make
sense when used to validate non-Finnish VAT-numbers.}
}
\value{
A logical vector that can be interpreted as follows:
 \describe{
  \item{NA}{The string has a correct format but an incorrect checksum-digit.}
  \item{FALSE}{Ths string has an incorrect format.}
  \item{TRUE}{The string appears to be of correct format.}
}
}
\description{
Functions for inferring if the input string is a valid Business ID, VAT
Number, E-Invoicing address or an ID Number. Rules obtained from
\url{http://tarkistusmerkit.teppovuori.fi/tarkmerk.htm}. Removes all
characters that can not be present in a valid string, such commas, dots
and white space.
}
