% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vat_regexes}
\alias{vat_regexes}
\title{Regexes for EU VAT Identification numbers}
\format{
A tibble with 28 rows and 3 columns
\describe{
  \item{code}{A country code.}
  \item{country}{The name of the country.}
  \item{format}{The regex for the VAT identification number}
}
}
\source{
\url{http://ec.europa.eu/taxation_customs/vies/faqvies.do#item_11}
}
\usage{
vat_regexes
}
\description{
A dataset containing regexes for VAT numbers (excluding whitespaces).
}
\keyword{datasets}
