% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\alias{valid_yt}
\alias{valid_ovt}
\alias{valid_vat}
\alias{valid_pic}
\title{Check if the string is a valid ID Number}
\usage{
valid_yt(x, require_checksum = TRUE)

valid_ovt(x, require_checksum = TRUE)

valid_vat(x, require_checksum = TRUE)

valid_pic(x, require_checksum = TRUE)
}
\arguments{
\item{x}{The input string(s).}

\item{require_checksum}{If \code{TRUE} (default), require that the checksum
digit matches in addition to the format of the string. Note, does not make
sense when used to validate non-Finnish VAT-numbers.}
}
\value{
A logical vector that can be interpreted as follows:
 \describe{
  \item{NA}{The string has a correct format but an incorrect checksum-digit.}
  \item{FALSE}{Ths string has an incorrect format.}
  \item{TRUE}{The string appears to be of correct format.}
}
}
\description{
Functions for inferring if the input string is a valid Business ID, VAT
Number, E-Invoicing address or a personal identification code (PIC).
Rules obtained from \url{http://tarkistusmerkit.teppovuori.fi/tarkmerk.htm}.
Removes all characters that can not occur in a valid string, such commas,
dots and white space.
}
